+++ lua/obsidian/completion/refs.lua
@@ -2,12 +2,6 @@ local util = require "obsidian.util"
 
 local M = {}
 
----@enum obsidian.completion.RefType
-M.RefType = {
-  Wiki = 1,
-  Markdown = 2,
-}
-
 ---Backtrack through a string to find the first occurrence of '[['.
 ---
 ---@param input string
@@ -20,8 +14,6 @@ local find_search_start = function(input)
       return nil
     elseif vim.startswith(substr, "[[") then
       return substr, string.sub(substr, 3)
-    elseif vim.startswith(substr, "[") and string.sub(input, i - 1, i - 1) ~= "[" then
-      return substr, string.sub(substr, 2)
     end
   end
   return nil
@@ -36,7 +28,6 @@ end
 ---@return string|? search_string
 ---@return integer|? insert_start
 ---@return integer|? insert_end
----@return obsidian.completion.RefType|? ref_type
 M.can_complete = function(request)
   local input, search = find_search_start(request.context.cursor_before_line)
   if input == nil or search == nil then
@@ -49,12 +40,7 @@ M.can_complete = function(request)
     local suffix = string.sub(request.context.cursor_after_line, 1, 2)
     local cursor_char = request.context.cursor.character
     local insert_end_offset = suffix == "]]" and 1 or -1
-    return true, search, cursor_char - vim.fn.strchars(input), cursor_char + 1 + insert_end_offset, M.RefType.Wiki
-  elseif vim.startswith(input, "[") then
-    local suffix = string.sub(request.context.cursor_after_line, 1, 1)
-    local cursor_char = request.context.cursor.character
-    local insert_end_offset = suffix == "]" and 0 or -1
-    return true, search, cursor_char - vim.fn.strchars(input), cursor_char + 1 + insert_end_offset, M.RefType.Markdown
+    return true, search, cursor_char - vim.fn.strchars(input), cursor_char + 1 + insert_end_offset
   else
     return false
   end