+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- New `link` module options: `link.style`, `link.format`.
+- `link.format` is fully obsidian app compatible, controlling how path is formatted
+  - `shortest`: shortest file path when possible.
+  - `relative`: relative path from the current note file.
+  - `absolute`: vault-relative absolute path.
+
+### Changed
+
+- Deprecated `preferred_link_style`, `wiki_link_func`, and `markdown_link_func` in favor of `link.style`.
+
 ## [v3.15.11](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.15.11) - 2026-03-05
 
 ### Added